helpdesk-agent/pyproject.toml
[project]
name = "helpdesk-agent"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.9",
    "anthropic>=0.49",
    "python-dotenv>=1.0",
]

[tool.uv]
package = false

[project.scripts]
helpdesk-agent = "src.main:main"
